Update: All Database Type Test Complete Fix: Implement MySQL type override for Float to Double due to lack of precision in Unit Tests. Fix: Implement SQLite type override for Sbyte, UInt, ULong, due to poor handling...

+++ DOLDatabase/Handlers/MySQLObjectDatabase.cs
@@ -108,7 +108,8 @@ namespace DOL.Database.Handlers
 			}
 			else if (bind.ValueType == typeof(float))
 			{
-				type = "FLOAT";
+				// Float Value have less precision than C# Single.
+				type = "DOUBLE";
 			}
 			else if (bind.ValueType == typeof(double))
 			{
